# internal/controller/api/api.spec.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "description": "Local Receptor-Controller API",
      "url": "/api/receptor-controller/v1/"
    }
  ],
  "info": {
    "description": "Receptor-Controller Service",
    "version": "1.0.0",
    "title": "Receptor-Controller"
  },
  "tags": [
    {
      "name": "api",
      "description": "RESTful service"
    }
  ],
  "paths": {
    "/job": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "Submit a job request to be routed to a customers environment",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "404": {
            "description": "No connection to the target receptor node"
          }
        }
      }
    },
    "/connection": {
      "get": {
        "tags": [
          "api"
        ],
        "summary": "Get a list of open connections",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/connection/status": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "Submit a job request to be routed to a customers environment",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionStatusResponse"
                }
              }
            }
          }
        }
      }
    },
    "/connection/ping": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "Send a ping request to a receptor node",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionPingResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-rh-identity",
        "description": "Base64-encoded JSON identity header provided by 3Scale. Contains an account number of the user issuing the request. Format of the JSON: {\"identity\": {\"account_number\": \"12345678\"}}"
      }
    },
    "schemas": {
      "JobResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "JobRequest": {
        "type": "object",
        "properties": {
          "account": {
            "type": "string"
          },
          "recipient": {
            "type": "string"
          },
          "payload": {
            "type": "object"
          },
          "directive": {
            "type": "string"
          }
        }
      },
      "ConnectionListResponse": {
        "type": "object",
        "properties": {
          "connections": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "account": {
                  "type": "string",
                  "example": "0000001"
                },
                "connections": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "example": "node-a"
                  }
                }
              }
            }
          }
        }
      },
      "ConnectionStatusRequest": {
        "type": "object",
        "properties": {
          "account": {
            "type": "string"
          },
          "node_id": {
            "type": "string"
          }
        }
      },
      "ConnectionStatusResponse": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ConnectionStatus"
          },
          "capabilities": {
            "type": "object"
          }
        }
      },
      "ConnectionPingResponse": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ConnectionStatus"
          },
          "payload": {
            "$ref": "#/components/schemas/Payload"
          }
        }
      },
      "Payload": {
        "type": "object",
        "properties": {
          "account": {
            "type": "string"
          },
          "sender": {
            "type": "string"
          },
          "message_type": {
            "type": "string"
          },
          "message_id": {
            "type": "string"
          },
          "payload": {
            "type": "object"
          },
          "code": {
            "type": "integer"
          },
          "in_response_to": {
            "type": "string"
          },
          "serial": {
            "type": "integer"
          }
        }
      },
      "ConnectionStatus": {
        "type": "string",
        "enum": [
          "connected",
          "disconnected"
        ]
      }
    }
  }
}